Re-split compound words that were joined without a separator when they appeared at the end of a line
Some compound words like "corn-like" are broken down at the dash when they appear at the end of a line the BJCP PDF. They also appear with the dash when they are not at the end of a line. It stands to reason that the dash is not a line-continuation mark for those and should be preserved in JSON format.